Extracted ElasticSearch_seeItemExistsInElasticsearchTest
