+++ tests/ElasticSearchTest.php
@@ -16,33 +16,6 @@ class ElasticSearchTest extends ElasticSearchTestCase
     /**
      * @test
      */
-    public function seeItemExistsInElasticsearchShouldCallExistsWithIndexNameOnClient()
-    {
-        $this->module->seeItemExistsInElasticsearch('index-name', null, null);
-        $this->client->shouldHaveReceived('exists')->with(m::subset(['index' => 'index-name']))->once();
-    }
-
-    /**
-     * @test
-     */
-    public function seeItemExistsInElasticsearchShouldCallExistsWithTypeOnClient()
-    {
-        $this->module->seeItemExistsInElasticsearch(null, 'document-type', null);
-        $this->client->shouldHaveReceived('exists')->with(m::subset(['type' => 'document-type']))->once();
-    }
-
-    /**
-     * @test
-     */
-    public function seeItemExistsInElasticsearchShouldCallExistsWithIdOnClient()
-    {
-        $this->module->seeItemExistsInElasticsearch(null, null, 'document-id');
-        $this->client->shouldHaveReceived('exists')->with(m::subset(['id' => 'document-id']))->once();
-    }
-
-    /**
-     * @test
-     */
     public function grabAnItemFromElasticsearchShouldCallSearchWithIndexOnClient()
     {
         $this->module->grabAnItemFromElasticsearch('index-name');
